+++ clevertap-android-sdk/src/main/java/com/clevertap/android/sdk/CleverTapInstanceConfig.java
@@ -301,7 +301,8 @@ public class CleverTapInstanceConfig implements Parcelable {
     }
 
     @SuppressWarnings({"unused"})
-    void setEnableUIEditor(boolean enableUIEditor) {
+    @RestrictTo(RestrictTo.Scope.LIBRARY)
+    public void setEnableUIEditor(boolean enableUIEditor) {
         this.enableUIEditor = enableUIEditor;
     }
 
@@ -311,7 +312,8 @@ public class CleverTapInstanceConfig implements Parcelable {
     }
 
     @SuppressWarnings("SameParameterValue")
-    void setEnableABTesting(boolean enableABTesting) {
+    @RestrictTo(RestrictTo.Scope.LIBRARY)
+    public void setEnableABTesting(boolean enableABTesting) {
         this.enableABTesting = enableABTesting;
     }