+++ clevertap-android-sdk/src/main/java/com/clevertap/android/sdk/Constants.java
@@ -40,6 +40,8 @@ public class Constants {
     static final String ICON_BASE_URL = "http://static.wizrocket.com/android/ico/";
     static final String NOTIFICATION_CLICKED_EVENT_NAME = "Notification Clicked";
     static final String NOTIFICATION_VIEWED_EVENT_NAME = "Notification Viewed";
+    static final String GEOFENCE_ENTERED_EVENT_NAME = "Geocluster Entered";
+    static final String GEOFENCE_EXITED_EVENT_NAME = "Geocluster Exited";
     static final String APP_LAUNCHED_EVENT = "App Launched";
     static final String ERROR_KEY = "wzrk_error";
     static final int PUSH_DELAY_MS = 1000;
@@ -65,6 +67,7 @@ public class Constants {
     static final String DISPLAY_UNIT_JSON_RESPONSE_KEY = "adUnit_notifs";
     static final String FEATURE_FLAG_JSON_RESPONSE_KEY = "ff_notifs";
     static final String REMOTE_CONFIG_FLAG_JSON_RESPONSE_KEY = "pc_notifs";
+    static final String GEOFENCES_JSON_RESPONSE_KEY = "geofences";
     static final String DISCARDED_EVENT_JSON_KEY = "d_e";
     static final String INAPP_MAX_DISPLAY_COUNT = "mdc";
     static final String INAPP_MAX_PER_SESSION = "imc";
@@ -113,7 +116,9 @@ public class Constants {
     static final String KEY_COUNTS_PER_INAPP = "counts_per_inapp";
     static final String INAPP_ID_IN_PAYLOAD = "ti";
     static final int LOCATION_PING_INTERVAL_IN_SECONDS = 10;
-    static final String[] SYSTEM_EVENTS = {NOTIFICATION_CLICKED_EVENT_NAME, NOTIFICATION_VIEWED_EVENT_NAME};
+    static final String[] SYSTEM_EVENTS = {NOTIFICATION_CLICKED_EVENT_NAME,
+            NOTIFICATION_VIEWED_EVENT_NAME, GEOFENCE_ENTERED_EVENT_NAME,
+            GEOFENCE_EXITED_EVENT_NAME};
     static final long DEFAULT_PUSH_TTL = 1000 * 60 * 60 * 24 * 4;
     static final String PF_JOB_ID = "pfjobid";
     static final int PING_FREQUENCY_VALUE = 240;
@@ -230,4 +235,6 @@ public class Constants {
     public static final String LOG_TAG_PRODUCT_CONFIG = "Product Config : ";
     public static final int FETCH_TYPE_PC = 0;
     public static final int FETCH_TYPE_FF = 1;
+
+    public static final String LOG_TAG_GEOFENCES = "Geofences : ";
 }